change tests to use jest
