+++ packages/graphql-anywhere/package.json
@@ -5,16 +5,12 @@
   "main": "./lib/src/index.js",
   "typings": "./lib/src/index.d.ts",
   "scripts": {
-    "pretest": "npm run compile",
-    "test": "npm run testonly --",
-    "posttest": "npm run lint",
+    "test": "jest",
+    "posttest": "npm run lint && npm run compile",
     "compile": "tsc",
     "watch": "tsc -w",
     "prepublishOnly": "npm run compile",
     "lint": "grunt tslint",
-    "coverage": "istanbul cover ./node_modules/mocha/bin/_mocha -- --reporter dot --full-trace lib/test/tests.js",
-    "postcoverage": "remap-istanbul --input coverage/coverage.json --type lcovonly --output coverage/lcov.info",
-    "testonly": "mocha --reporter spec --full-trace lib/test/tests.js",
     "compile:browser": "browserify ./lib/src/index.js -o=./lib/index.js --i apollo-utilities && npm run minify:browser",
     "minify:browser": "uglifyjs -c -m -o ./lib/index.min.js -- ./lib/index.js",
     "filesize": "npm run compile:browser"
@@ -38,32 +34,34 @@
     "apollo-utilities": "*"
   },
   "devDependencies": {
-    "@types/chai": "3.5.2",
     "@types/graphql": "~0.9.0",
     "@types/isomorphic-fetch": "0.0.34",
+    "@types/jest": "^20.0.7",
     "@types/lodash": "^4.14.37",
-    "@types/mocha": "^2.2.32",
     "@types/node": "^8.0.0",
     "@types/react": "^15.0.14",
     "@types/react-dom": "^15.5.0",
     "async": "^2.0.0",
     "browserify": "^14.4.0",
-    "chai": "^4.1.0",
-    "chai-as-promised": "^7.0.0",
     "es6-promise": "^4.0.5",
     "graphql": "^0.10.0",
     "graphql-tag": "^2.0.0",
     "grunt": "1.0.1",
     "grunt-tslint": "5.0.0",
     "isomorphic-fetch": "^2.2.1",
-    "istanbul": "^0.4.5",
-    "mocha": "^3.0.0",
+    "jest": "^20.0.4",
     "react": "^15.3.1",
     "react-dom": "^15.3.1",
-    "remap-istanbul": "^0.8.0",
-    "source-map-support": "^0.4.0",
+    "ts-jest": "^20.0.10",
     "tslint": "^5.5.0",
     "typescript": "^2.2.1",
     "uglify": "^0.1.5"
+  },
+  "jest": {
+    "transform": {
+      ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+    },
+    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
+    "moduleFileExtensions": ["ts", "tsx", "js", "json"]
   }
 }
